Added roof. Corrected collider positions
The roof is still incorrect in some instances. Currently I use the bounds of the mesh, which is extended for the window sills I created in the modeds. This makes the roof spill over the wall.
Doubt settles in again. This project requires very precise vector math regarding positions and dimensions.